requirements.txt
flask>=2.3.0
requests>=2.31.0
python-dotenv>=1.0.0
agora-token-builder>=1.0.0
flask-cors>=4.0.0
pytz>=2024.1
